+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/ClientCommon.cs
@@ -128,15 +128,5 @@ namespace Azure.AI.MetricsAdvisor
                 throw new ArgumentException(UnexpectedHeaderFormat);
             }
         }
-
-        internal static HookInfoPatch GetPatchModel(AlertingHook hook)
-        {
-            return hook switch
-            {
-                EmailHook h => new EmailHookInfoPatch() { HookName = h.Name, Description = h.Description, ExternalLink = h.ExternalLink, HookParameter = h.HookParameter, Admins = h.Administrators },
-                WebHook h => new WebhookHookInfoPatch() { HookName = h.Name, Description = h.Description, ExternalLink = h.ExternalLink, HookParameter = h.HookParameter, Admins = h.Administrators },
-                _ => throw new InvalidOperationException("Unknown AlertingHook type.")
-            };
-        }
     }
 }
