+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/Hook/AlertingHook.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 using Azure.Core;
 
@@ -32,5 +33,15 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// </summary>
         [CodeGenMember("Admins")]
         public IReadOnlyList<string> Administrators { get; }
+
+        internal static HookInfoPatch GetPatchModel(AlertingHook hook)
+        {
+            return hook switch
+            {
+                EmailHook h => new EmailHookInfoPatch() { HookName = h.Name, Description = h.Description, ExternalLink = h.ExternalLink, HookParameter = h.HookParameter, Admins = h.Administrators },
+                WebHook h => new WebhookHookInfoPatch() { HookName = h.Name, Description = h.Description, ExternalLink = h.ExternalLink, HookParameter = h.HookParameter, Admins = h.Administrators },
+                _ => throw new InvalidOperationException("Unknown AlertingHook type.")
+            };
+        }
     }
 }
